requirements.txt
pypdf
pymupdf
pdfplumber
python-docx
pytesseract
pillow
pandas
openpyxl
beautifulsoup4
lxml
python-dotenv
tqdm
openai
anthropic
google-genai
groq
